[feat 175] add Views system for usertype
