+++ app/client/src/app/(main)/(wallet-connection-needed)/my-collections/page.tsx
@@ -1,9 +1,11 @@
 'use client'
 
+import { Header } from "@/components/Headers/Header";
+
 export default function Dashboard() {
   return (
     <div className="">
-        My Collections
+        <Header title="Collections" hasCreateButton={true} />
     </div>
   );
 }
