+++ app/client/src/utils/AddressFormat.ts
@@ -1,4 +1,5 @@
-function shortAddress(address: string): string {
+function shortAddress(address?: string): string {
+    if(!address) return ""
     return `${address.slice(0, 6)}...${address.slice(-4)}`;
 }
